Move dataset states to utilities, remove unused backbone dataset models
