+++ client/src/composables/monitorUploadedHistoryItems.ts
@@ -4,7 +4,7 @@ import { computed, type Ref } from "vue";
 import { type HDASummary, type HistoryItemSummary, isHDA } from "@/api";
 import type { ComponentColor } from "@/components/BaseComponents/componentVariants";
 import type { UploadItem } from "@/components/Upload/model";
-import STATES from "@/mvc/dataset/states";
+import STATES from "@/utils/datasetStates";
 import { useHistoryItemsStore } from "@/stores/historyItemsStore";
 
 const REFER_TO_HISTORY_MSG = "Refer to the history panel to view dataset state.";
